feat: complete Bubble Transformer V4 real experiment with Qwen3-0.6B
- Execute epsilon sweep finding sweet spot at ε=0.001 (606.7 effective rank, 0.044 concentration ratio)
- Run dual-head tension sweep showing single-head is optimal
- Fix float16 overflow in embedding extraction (convert to float32)
- Apply outlier clipping (percentiles 0.5-99.5) for stable metrics
- Update config for Qwen3-0.6B (d_model=1024, num_heads=16)
- Add comprehensive documentation in docs/testreales/ (8 files, 68KB)
- Generate 14 visualization plots showing concentration patterns
- Add V4 core implementation and test suite (44 tests passing)
- Include arXiv paper draft and SVG diagram

+++ experiments/extract_embeddings.py
@@ -135,7 +135,8 @@ class QwenEmbeddingExtractor:
         else:
             hidden_states = output
         # Move to CPU immediately to save VRAM
-        self.layer_outputs[layer_idx] = hidden_states.cpu()
+        # Convert to float32 to prevent overflow from quantization artifacts
+        self.layer_outputs[layer_idx] = hidden_states.float().cpu()
 
     def _get_memory_usage(self) -> str:
         """Return current GPU memory usage string."""
@@ -169,11 +170,11 @@ class QwenEmbeddingExtractor:
         with torch.no_grad():
             self.model(input_ids=input_ids, attention_mask=attention_mask)
 
-        # Convert to numpy
+        # Convert to numpy (float32 for numerical stability)
         result = {}
         for layer_idx, hidden in self.layer_outputs.items():
-            # hidden: [B, N, D] — already on CPU from hook
-            result[layer_idx] = hidden.numpy()
+            # hidden: [B, N, D] — already on CPU from hook, float32
+            result[layer_idx] = hidden.numpy().astype(np.float32)
 
         return result
